Update guidance on situations where there is no polkit authority
Now that GDBusProxy does something reasonable for a masked systemd
service, see

 https://bugzilla.gnome.org/show_bug.cgi?id=677718

construction of the PolkitAuthority object does not fail anymore. That
doesn't mean the authority is available, though, so mention that users
should check the result of the CheckAuthorization() call as well. Or
in the case of PolkitAuthority, that the error is not a POLKIT_ERROR.

This is actually a nice feature, it means that if you unmask
polkit.service then mechanisms using PolkitAuthority will start using
it without a restart.

Signed-off-by: David Zeuthen <zeuthen@gmail.com>