authority: Avoid cookie wrapping by using u64 counter
Tavis noted that it'd be possible with a 32 bit counter for someone to
cause the cookie to wrap by creating Authentication requests in a
loop.

My current belief is that the most effect this could have would be
that in the case where there are two active sessions, Mallory could
cause Alice's authorization checks to spuriously fail if she managed
to create a request with a duplicate cookie.

Something important to note here is that wrapping of signed integers
is undefined behavior in C, so we definitely want to fix that.  The
cookie counter is now unsigned.

I chose to extend the counter to 64 bits, add the polkit process start
time as well as an additional random 32 bits.  At this point we're not
too far from just having a UUID, but one thing I do like about the
counter approach is that it is process scoped, we don't actually need
something "universally" unique.

Reported-by: Tavis Ormandy <taviso@google.com>
Signed-off-by: Colin Walters <walters@redhat.com>