configure: enable elogind support in PolicyKit
Currently configure.ac allows the usage of either libsystemd-login or
ConsoleKit for session tracking.

The elogind project (https://github.com/wingo/elogind) provides
systemd-logind as a stand-alone daemon.  To allow users of systems not
run by the full systemd suite, for whatever reason there might be, to at
least be able to not have to use ConsoleKit for session tracking, an
integration of elogind usage into PolicyKit would be very helpfull.

As elogind provides the same interface as systemd-logind, integration is
easy.

This commit changes configure.ac to look for elogind in the same way
it looks for systemd-logind.

https://bugs.freedesktop.org/show_bug.cgi?id=98688