text listener: Add properties
Add properties to turn off the use of color, to enable
the use of the alternate screen buffer, and to add a delay
before switching back from the alernate screen buffer.

This makes the text listener usable in situations where
just printing the text would disturb formatting on the
screen.