polkitd: add logging for successfully loaded rules
Log each successful load operation of a rule file including the name
of the loaded file. Getting the list of active rule files from the
logs is an important information in a support case and should not need
additional tracing/debugging.
---
Built and tested on top of 4b7a5c3. The info is logged during polkitd
startup and with changes of rule files, also in case '--no-debug' is set.
It might be a good idea to add a log_level to the argument list of
polkit_backend_authority_log() but this would affect >60 callers and is
out of scope of this patch.