Enforce that calling CheckAuthorization() with pidfd also is passed a uid
When unix-process is used and a uid is not passed, it will be derived from
the running process. But this is racy, as the identity can change over
time (e.g.: setuid binaries can be exec'ed). If the caller can securely
fetch the pidfd of a process (e.g.: via SO_PEERPIDFD), then it can also
securely fetch the uid (e.g.: via SO_PEERCRED), so enforce that they are
passed together to avoid possible races, in case a polkit rule performs
authorization based on the uid.

Follow-up for 374a628038474a45f