test: temporarily disable test-polkitbackendjsauthority under ASan
The test, in its current form, is incompatible with ASan, since it
overwrites $LD_PRELOAD and thus preventing ASan to work properly. This
can be worked around using `verify_asan_link_order=0`, but this only
shifts the problem down the road, as with that ASan is able to run but
produces false positive reports since it can't properly incercept calls
made to libduktape.