Implement configuration reloading on SIGHUP and sd_notify support
Add support for active reloading of config files via SIGHUP.
Also integrate with sd_notify when building with libsystemd, and
set Type=notify-reload in the systemd unit, so that a 'systemctl
reload polkit' works out of the box.
This is especially useful when booting a read-only image, which
gets configuration updates via overmounting with another read-only
image.