Revert "Implement LogControl1 protocol for dynamic log level changes"
After a long discussion, the polkit team decided to revert the
implementation of LogControl1 in polkit for the following reasons:
1. The authentication/authorization part of changing LogLevel needs to
   be implemented. Without it, any user can change log level via busctl.
   Issue: !507
2. Issue: !506
After all these issues are addressed, re-integration of the protocol can
be later reconsidered.
LogControl1 is not widely used so far, therefore no major damage will
be caused by postponing the integration of the protocol within polkit.

This reverts commit 79d84d0dd0dd7d16c616555acf8cd14862d02baf.