test: run dfuzzer against polkit's D-Bus interface
Let's run dfuzzer to fuzz polkit's own D-Bus interface to catch issues
like [0][1][2] early.

The test implementation is a _bit_ convoluted, caused by the fact that
we have to rebuild polkit with Address and Undefined sanitizers to make
dfuzzer considerably more effective. This also involves running the just
built polkit in a lightweight container so we don't mess with the
already installed polkit (which could affect tests running after this
one that don't expect sanitized polkit). All this extra code can be
potentially dropped once we're able to do multiple Packit builds in one
PR (currently in an RFC stage).

Resolves: #515

[0] https://github.com/polkit-org/polkit/issues/506
[1] https://github.com/polkit-org/polkit/commit/b709b693346712bf8e8d8e8c77f71b7918fa367d
[2] https://github.com/polkit-org/polkit/commit/5cce296a8d60a3395c1a372377bdf8f01ed93872