polkitpermission: add a paranoia check for the error object
The error object in this case should always be allocated if
polkit_authority_check_authorization_finish() returns NULL, so let's
convey this message to static analyzers as well.

Another potential "solution" for this could be checking the error object
instead of the result one when returning from
polkit_authority_check_authorization_finish() (like it's done in the
respective example - src/examples/cancel.c), but we already have a
precedent of using g_assert() on the error object for this in our
codebase.

This should address following warning from GCC analyzer:

[75/89] Linking target src/polkit/libpolkit-gobject-1.so.0.0.0
../src/polkit/polkitpermission.c: In function ‘changed_check_cb’:
../src/polkit/polkitpermission.c:485:7: warning: dereference of NULL ‘error’ [CWE-476] [-Wanalyzer-null-dereference]
  485 |       g_warning ("Error checking authorization for action id %s: %s",
      |       ^