packit: build Packit with LTO in OpenScanHub
GCC analyzer in some cases depends on information provided by LTO to
generate accurate results, but the LTO support in GCC analyzer is still
experimental, so it's not enabled by default. In Polkit, however, this
seems to work pretty well, so let's tell OSH to build our code with LTO
when running GCC analyzer.