test: rewrite tests for realpath() shenanigans
The second test tried to strace a setuid binary as an unprivileged user
which will never work, but the fact was hidden by missing strace on the
test machine. Let's rewrite the tests to fix this, and check a couple
more cases.

Follow-up for 9aa43e089d870a8ee695e625237c5b731b250678.