PolkitUnixProcess: Don't set pidfd in pid setter during construction
All G_PARAM_CONSTRUCT properties are set during construction in an
unspecified order. For this reason a property setter shouldn't modify
fields that are tied to other properties during the initial object
construction, or a change in the unspecified property initialization
order will result in unpredictable behavior changes.

Fix this by checking if the object is still being constructed, and if
so, assume that the pidfd will be set later by `->constructed()` and
skip setting it in the pid property setter.