PolkitUnixProcess: Fix hashing with pidfd
If the current PolkitUnixProcess object is associated with a pidfd, it's
possible for the value of polkit_unix_process_get_pid to unexpectedly
change even when the object itself is not modified.

Fix this by just using the creation time for hashing if the object isn't
associated with a pidfd.

This fixes another pidfd leak issue with registering authentication
agents.