build: Disable deprecation warnings for GLib versions we don’t depend on
Use the version bounding macros provided by GLib to expose the
configure-time dependency version to the preprocessor, so it knows
not to enable deprecation warnings for GLib APIs deprecated in versions
of GLib newer than the one polkit depends on.

This prevents a load of build-time warnings about using
`g_simple_async_result_*()` APIs. Those APIs were deprecated in GLib
2.46, but polkit only depends on 2.44.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>