Early destroy of main context during source destroy causes deadlock
It seems that main context is being destroyed in another thread while
GAsyncResult instance is destroying g_source attached to the context.
A reference to the context seems to be missing, so it is destroyed early
during thread race.
This is rather a temporary fix until GAsyncResult is rewritten to GTask
(with better documentation, hopefully), which might also clean the build
log of deprecation warnings.

Fixes !568