g_source_destroy as DestroyNotify callback runs into deadlock with MainContext destroy (#573)
It seems that when g_source_destroy as callback is given explicitly, it runs into a deadlock with associated GMainContext which already is destroying the Source, putting a rw lock on it.

Two threads, both destroying a source attached to a single context. One destroys it because of flatlined refs, other because g_source_destroy is called upon the source as destroy callback.
Feedback welcome, the Glib documentation is... well, entangled (/euphemism).

Fixes !568