polkitpermission: Fix a leak of systemd session state data (#571)
* build: Bump GLib dependency to ≥2.44.0

polkit already implicitly depended on this through its use of
`g_steal_pointer()`, which was introduced in GLib 2.44.

2.44.0 was released in 2015 and is widely available.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

* polkitpermission: Fix a leak of systemd session state data

If the string comparison shows the two strings are equal, the value
returned from `get_session_state()` is currently leaked:
```
Direct leak of 2752 byte(s) in 43 object(s) allocated from:
    #0 0x7fc750ce5e2b in realloc.part.0 (/lib64/libasan.so.8+0xe5e2b) (BuildId: 7f1aa7e2e600e8c9d54ce6e3d36f3d31bfe7949a)
    #1 0x7fc74df8e343 in greedy_realloc (/lib64/libsystemd.so.0+0xa1343) (BuildId: c252dd7627af268244ad25e5141e0a280c0d68d2)
    #2 0x7fc74df8ec4a in parse_env_filev.constprop.0 (/lib64/libsystemd.so.0+0xa1c4a) (BuildId: c252dd7627af268244ad25e5141e0a280c0d68d2)
    #3 0x7fc74df92635 in parse_env_file_sentinel.constprop.0 (/lib64/libsystemd.so.0+0xa5635) (BuildId: c252dd7627af268244ad25e5141e0a280c0d68d2)
    #4 0x7fc74df6bd29 in sd_session_get_state (/lib64/libsystemd.so.0+0x7ed29) (BuildId: c252dd7627af268244ad25e5141e0a280c0d68d2)
    #5 0x7fc75049fd3d in get_session_state (/lib64/libpolkit-gobject-1.so.0+0x3d3d) (BuildId: f8009e26e007bc067476989a7baa2584b23fb09e)
    #6 0x7fc7504a9202 in on_sessions_changed (/lib64/libpolkit-gobject-1.so.0+0xd202) (BuildId: f8009e26e007bc067476989a7baa2584b23fb09e)
```

Fix that by stealing the new value if it’s moved into the struct member,
and otherwise freeing it.

This whole function could be simplified significantly by using
`g_set_str()`, but that’s only available in GLib 2.76, which polkit
does not currently depend on.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

* build: Disable deprecation warnings for GLib versions we don’t depend on

Use the version bounding macros provided by GLib to expose the
configure-time dependency version to the preprocessor, so it knows
not to enable deprecation warnings for GLib APIs deprecated in versions
of GLib newer than the one polkit depends on.

This prevents a load of build-time warnings about using
`g_simple_async_result_*()` APIs. Those APIs were deprecated in GLib
2.46, but polkit only depends on 2.44.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

---------

Signed-off-by: Philip Withnall <pwithnall@gnome.org>